added a config file and read_df fxn
thread_df fxn reads in df based on file extension in filepath. the config file (as shown in test.py) allows looping through different files (for example if you have a staff and client data in separate files). WIP submission to gen3. May want to output a file with multiindex with a second level that specifies node (ie variable category).